add a MovingAverage class
